+++ src/types/index.ts
@@ -1,6 +1,7 @@
 export * from "./changesets";
 export * from "./features";
 export * from "./general";
+export * from "./messages";
 export * from "./notes";
 export * from "./osmPatch";
 export * from "./user";