+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## 2.2.0 (2024-09-16)
+
 - Added 3 new functions for the preferences API
 - Added 5 new functions for the new messaging API
 - Added a new option `bbox` to `getNotesForQuery`